Render maps as tables
UI code is getting some structure. I need to move it out of main
though but since I'm still playing around with the whole Bubble tea
thing it's ok to keep in main for now.

Initial idea: render both vectors and maps as tables.
I've coded up nested maps and nested vectors as tables as well but I
will need to research how to better do it. I don't want to actually
show the nested tables right away but have them be extendable on key
press (Enter most probably). But this is more as a note to my future
self.